+++ cogeo_mosaic/backends/dynamodb.py
@@ -167,16 +167,17 @@ class DynamoDBBackend(BaseBackend):
             self.mosaic_def.maxzoom,
             quadkey_zoom=self.quadkey_zoom,
             tilematrixset=self.mosaic_def.tilematrixset,
-            asset_type=self.mosaic_def.asset_type,
-            asset_prefix=self.mosaic_def.asset_prefix,
-            data_type=self.mosaic_def.data_type,
-            colormap=self.mosaic_def.colormap,
-            layers=self.mosaic_def.layers,
             quiet=quiet,
             **kwargs,
         )
 
         bounds = bbox_union(new_mosaic.bounds, self.mosaic_def.bounds)
+
+        if self.mosaic_def.mosaicjson != new_mosaic.mosaicjson:
+            warnings.warn(
+                f"Updating `mosaicjson` version from {self.mosaic_def.mosaicjson} to {new_mosaic.mosaicjson}"
+            )
+            self.mosaic_def.mosaicjson = new_mosaic.mosaicjson
 
         self.mosaic_def._increase_version()
         self.mosaic_def.bounds = bounds